Renamed directories for case-insenstive systems
